World 🌍
This update introduces the camera, the world and the worldgen systems. This is a massive overhaul! On start, players will be randomly placed around the world, after the host generates the world with the configuration menu. (Improvements to come to this later)

The worldgen system uses noise, cells, chunks, materials and biomes to create a realistic environment. Terrain materials are chosen in cells, which is then upscaled for management across chunks. These chunks are also stored in biomes, which are equivalent to heightmaps essentially.

Every pixel generated in the world has information. Stored as UINTs we know the height, color, and all other information exposed through the material.

CRTL+G = Show Chunks/Biomes/Tooltip
CTRL+' = Regen world during game

I have not tested the above regen in multiplayer.

The camera is the old canvas. Eventually this will now be able to be resized, as all positional calculations and rendering are now world relative, with some camera culling. This camera is more like a viewport.

Fixed all uppercasing strings and started moving CONFIG defs from the general flat file to their own [script]config.ts.

Decals and particles all rendered camera and world relative, and still persistent during games.

Changed player quips on hit from sfx mixer to voice.